backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
openai
pillow
pypdf2
requests
pdf2image